requirements.txt
beautifulsoup4==4.12.2
boto3==1.24.28
decorator==5.1.1
gspread==5.10.0
json5==0.9.6
numpy==1.24.3
os==0.1.0
pandas==1.5.3
python-dotenv==0.19.1
plotly==5.9.0
requests==2.31.0